requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appdirs==1.4.3            # via black
atomicwrites==1.1.5       # via pytest
attrs==18.1.0             # via black, pytest
black==18.6b4
click==6.7                # via black, pip-tools
coverage==4.5.1           # via pytest-cov
django-environ==0.4.5
django==2.1
first==2.0.1              # via pip-tools
iso8601==0.1.12           # via lepo
jsonschema==2.6.0         # via lepo
lepo==0.2.0
marshmallow==2.15.4       # via lepo
more-itertools==4.3.0     # via pytest
pip-tools==2.0.2
pluggy==0.7.1             # via pytest
py==1.5.4                 # via pytest
pytest-cov==2.5.1
pytest-django==3.3.3
pytest==3.7.1
pytz==2018.5              # via django
six==1.11.0               # via more-itertools, pip-tools, pytest
toml==0.9.4               # via black
ulid2==0.1.1
whitenoise==3.3.1
